Stencil and Mask Operators
This change adds a material switching system!  This allows for the appearance
of a model to be decoupled from the surface of the model.  This prevents false
zero-crossing meshing artifacts when used correctly, which in turn ensures
surface normals will be more correct.

The appearance is still volumetric, so the stencil and mask nodes are useful
for defining the interior structure of a model as well!