New Visibility Mode and a Crash Fix
This change introduces the "imminent" visibility mode for models that are
hidden but expected to be marked visible soon.  This allows for vertex colors
to be generated in advance.  This change also adjusts the "invisible" mode to
work such that vertex colors are not operated so as to yield the thread pool.

This change also attempts to fix a rare crash bug I encountered while
implementing this.  I believe it to be fixed now, but it was a null access
violation upon the Instance->Colors array in ShaderTask::Run.  This looked to
be possible because the dirty flag initializes to true (somewhat pointlessly),
so it was possible for ShaderTask::Run to hit it's inner branch as soon as
Painter->MeshReady flipped to true.  I noticed that Instance->CameraOrigin was
zero still for one of the two occurrances of this crash I ran into, which
suggests that it happened early enough that the renderer never hit the draw
call for the model.  The model was not visible both times, but the draw method
normally early-outs in these situations, after setting some atomics.

I'm planning on reworking how this data is modeled and synchronized after I
tear out the old renderer, so this is probably fine for now.