Conditional Lua GC After Callbacks, Change RNG Type
The conditional garbage collection run is meant to prevent LuaModel instances
from appearing to disappear "at random" when their references have expired.

The change from std::default_random_engine to std::mt19937 for Lua's RNG is
unrelated in purpose, but serves to ensure that the RNG behavior is consistent
between compilers and platforms.