Nicer Random Seed Randomization
Call `random_seed` with no args, and it will produce a random seed that is a
combination of whatever std::random_device produces (non-deterministic on some
systems), and the number of nanoseconds since the unix epoc (millisecond
precision on some systems).

The seed used is now returned by `random_seed` in all cases, to make it easier
to reply seeds, should a game want that.