New Transform Type to Replace TransformMachine
The new Transform type is meant to eventually replace the internals of the old
TransformMachine.  Currently it is only replaced for model instances.

This new type represents transforms as a quaternion rotation, a vector offset,
and a scalar scale.  Unlike TransformMachine, this does not require any special
lazy evaluation step.

This change also exposes a uniform scale function to model instances in the Lua
API.  This is for testing the scale functionality.  I'd like to eventually
support non-uniform scaling for model instances.

The TransformMachine declaration has been moved back into sdf_evaluator.cpp for
now, and will be replaced later when I rewrite the sdf evaluator.