Fixed Blend Set Function Thresholds When Scale is Applied
This bug only affected regular evaluation through the node graph.  Evaluation
via the interpreter was working correctly.

THIS HAS BEEN IRRITATING ME ALL YEAR

MAYBE LONGER I DON'T KNOW

Anyways, the problem turned out to be that SetNode was written with the
assumption that the threshold would never change.  This assumption was violated
when I implemented a scale verb, which requires scaling the threshold.  As this
is always done on a copy of the SDFNode tree, it was doing this weird
semi-imutable nonsense.

I'd like to get rid of this obnoxious mixin pattern eventually, now that I the
evaluator doesn't need to share math implementations with shader sources.