Removed SDF Interpreter "Push" Instruction
The stack behavior is intrinsic to the program, and therefor no special push
instruction was ever required.  I also switched out the stack implementation to
remove the need to explicitly manage the stack pointer.  It wasn't causing any
problems as far as I know, but it was just asking for trouble anyway.