Configurable Convergence Sequence
The meshing system can now reorder vertices in memory, which is useful for
manipulating the convergence artifact that occurs when the threadpool can't
keep up with the shading work.

Tangerine now defaults to the shuffle strategy somewhat arbitrarily, on the
basis that rapidly rotating objects look a bit better with it.  However, this
strategy greatly reduces vertex processing performance on the GPU by reducing
the vertex cache hit rate.  Games with more complex scenes will want to turn
this off.

Like the meshing density, you can turn this on and off between instancing
meshes to set the mode for the painter.  If a painter already exists that can
be shared by a new mesh instance, the current global convergence hint will be
ignored.

Here's the supported variations of the new API:

set_convergence_hint("serendipity")
set_convergence_hint("fastest")
set_convergence_hint("shuffle")
set_convergence_hint("diffuse")

The "serendipity" mode performs no vertex reordering, which means the vertex
ordering will be populated in whatever order the meshing algorithm chooses to
populate the vertex arrays.  The "shuffle" mode shuffles the vertex order in
memory and rewrites the index buffer array accordingly.  The "fastest" and
"diffuse" modes are aliases to "serendipity" and "shuffle" respectively.