Rework Repaint Mechanism to Not Rely on Clock
Previously this used std::chrono::high_resolution_clock on the assumptions
that it would be coherent accross threads, and that it advances while threads
are sleeping.  However, these assumptions appear to not be guaranteed:

https://groups.google.com/a/isocpp.org/g/std-discussion/c/D_S47zTnNmw

This new mechanism uses a simple fence counter to accomplish the same thing.
The old implementation was probably broken, as scenes are converging faster
now.

It may be a good idea to replace std::high_resolution_clock entirely, and only
allow the main thread to produce the timepoints to be shared between threads.