MSAA Support
ES2 had MSAA before this change, but now it is supported for all targets!

Tangerine will now attempt to allocate render targets with 8 samples per pixel
by default, and will gracefully fallback to lower counts or by disabling MSAA
when the requested sample count is not available.

This change also adds the "--msaa" switch, which lets the user override the
target sample count.