Clarify Parallel Task Type Names
This change renames the parallel task structs to not imply that they are only
for meshing.

I plan on revising this mechanism later to add a "baton pass" mechanism so that
chained tasks can pass a struct of intermediaries along, which should make this
more immediately useful outside the meshing systems.