Decouple Parallel Task Machinery from Meshing Intermediaries
This decouples ParallelDomainTaskChain and related machinery from the SDF
octree and painter types in favor of an "intermediary" struct who's lifetime
is managed by the task chain.

With this change the lifetimes of the the octree and painter objects durring
meshing now is held through the meshing process, as the parallel task machinery
now requires all of the intermediary data to have a lifetime that at least
exceeds the task chain.

While this is intended to make it easier to introduce parallelized work
elsewhere in tangerine, the more immediate use is to make it easier to specify
custom meshing scratch structs for the different meshing systems, and not have
SDFModel be aware of any of them.