Threadsafe Color Updates And Better Throughput Regulation
This prevents the weird random flickering black splats that may have been
torn reads or something.  This new system now has the update parallel-for
communicate updates in batches back to the main thread using a thread
safe queue.

The update thread wont run the next parallel-for until the main thread
has finished processing the last one, which is a nice feature for
throttling updates when they're not really needed (despite being continous),
and preventing the updates from running ahead of the main thread and
causing stalls.

This isn't perfect, though.  At higher loads convergence artifacts worsen,
but update pace can't increase to compensate.  This is because processing
the updates is timeboxed so the main thread can't miss vblanks by having
to process too many updates.  This is arguably a feature, or at least is
a problem that can't be mediated without being able to update vertex buffers
from a thread pool.