Better Splat Mesh Generator
This version generates splats as a set of concentric vertex loops
instead of a deformed vertex grid, and is generally able to produce
a better vertex distribution with a lower vertex count.

With some tuning, I can get 1 million splats to render at 60 hz on
my laptop, which is substantially more than is needed for this
particular demo.